+++ src/main/java/com/yht/exerciseassist/domain/emailCode/service/EmailService.java
@@ -14,11 +14,8 @@ import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpStatus;
 import org.springframework.mail.MailException;
 import org.springframework.mail.javamail.JavaMailSender;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-
-import java.time.LocalDateTime;
 
 @Service
 @Transactional
@@ -69,12 +66,5 @@ public class EmailService {
         message.setFrom(new InternetAddress("yhthealthassist@gmail.com", "YouHaveTo"));//보내는 사람
 
         return message;
-    }
-
-    @Scheduled(cron = "0 */1 * * * *")
-    public void deleteOldMailCode() {
-        LocalDateTime fiveMinutesAgo = LocalDateTime.now().minusMinutes(5);
-        emailCodeRepository.deleteByCreatedAtBefore(fiveMinutesAgo);
-        log.info("5분 지난 인증 코드 삭제");
     }
 }
