+++ src/main/java/com/yht/exerciseassist/domain/member/service/MemberService.java
@@ -23,7 +23,6 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.http.HttpStatus;
-import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.security.authentication.BadCredentialsException;
 import org.springframework.security.authentication.InternalAuthenticationServiceException;
 import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
@@ -38,7 +37,6 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 
 import java.io.IOException;
-import java.time.LocalDate;
 import java.time.LocalDateTime;
 import java.time.format.DateTimeFormatter;
 import java.util.Objects;
@@ -223,10 +221,10 @@ public class MemberService implements UserDetailsService {
         }
     }
 
-    @Scheduled(cron = "0 0 4 * * *")
-    public void deleteOldMemberData() {
-        String minusMonths = LocalDate.parse(LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM-dd"))).toString();
-        memberRepository.deleteByCreatedAtBefore(minusMonths);
-        log.info("탈퇴한지 30일 지난 멤버 데이터 삭제");
-    }
+//    @Scheduled(cron = "0 0 4 * * *")
+//    public void deleteOldMemberData() {
+//        String minusMonths = LocalDate.parse(LocalDateTime.now().minusMonths(1).format(DateTimeFormatter.ofPattern("yyyy-MM-dd"))).toString();
+//        memberRepository.deleteByCreatedAtBefore(minusMonths);
+//        log.info("탈퇴한지 30일 지난 멤버 데이터 삭제");
+//    }
 }
